requirements.txt
Flask==2.3.3
gunicorn==21.2.0
joblib==1.3.2
scikit-learn==1.6.1
numpy==2.0.0
pandas==2.1.4
python-dotenv==1.0.1
